# docs/swagger.yaml
definitions:
...
  models.Comment:
    properties:
      content:
        type: string
      review_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.Feature:
    properties:
      details:
        type: string
      name:
        type: string
    type: object
...
  models.Phone:
    properties:
      brand:
        type: string
      features:
        items:
          $ref: '#/definitions/models.Feature'
        type: array
      name:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
  models.PhoneRequest:
    properties:
      brand:
        type: string
      name:
        type: string
    required:
    - brand
    - name
    type: object
  models.Profile:
    properties:
      bio:
        type: string
      full_name:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Review:
    properties:
      content:
        type: string
      phone_id:
        type: integer
      rating:
        type: integer
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      comments:
        $ref: '#/definitions/models.Comment'
      password:
        type: string
      profile:
        $ref: '#/definitions/models.Profile'
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /phones:
    get:
      consumes:
      - application/json
      description: Get all phones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Phone'
            type: array
      summary: Get all phones
      tags:
      - phones
    post:
      consumes:
      - application/json
      description: Create a new phone
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone
        in: body
        name: phone
        required: true
        schema:
          $ref: '#/definitions/models.PhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new phone
      tags:
      - phones
  /phones/{phone_id}:
    delete:
      consumes:
      - application/json
      description: Delete a phone by ID
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Phone deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a phone by ID
      tags:
      - phones
    get:
      consumes:
      - application/json
      description: Get a phone by ID
      parameters:
      - description: Phone ID
        in: path
        name: phone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a phone by ID
      tags:
      - phones
    put:
      consumes:
      - application/json
      description: Update a phone
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phone_id
        required: true
        type: integer
      - description: Phone
        in: body
        name: phone
        required: true
        schema:
          $ref: '#/definitions/models.PhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      security:
      - ApiKeyAuth: []
      summary: Update a phone
      tags:
      - phones
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Create a new review
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - ApiKeyAuth: []
      summary: Create a new review
      tags:
      - reviews
  /reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a review
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a review
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Update a review
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - ApiKeyAuth: []
      summary: Update a review
      tags:
      - reviews
  /reviews/{phone_id}:
    get:
      consumes:
      - application/json
      description: Get reviews by phone ID
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get reviews by phone ID
      tags:
      - reviews
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - users
  /users/{id}/profile:
    put:
      consumes:
      - application/json
      description: Update user profile
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - users
swagger: "2.0"